setsockopt: intercept SO_TIMESTAMPING
Intercept setsockopt `SO_TIMESTAMPING` requests to rewrite hardware
timestamping requests to requests for software timestamps.

The datapath implementation is to copy in the scm_timestamping csmg
that hold timestamps, the value of ts[0] (which holds software
timestamps) to ts[2] (which holds raw hardware timestamps).

As a result, it is trivial to support requests for both hw and sw
timestamps.

The selftest needs the follow-up recvmsg timestamp conversion to pass.