recv: support MSG_ONEPKT
Define this flag.

No changes are needed in the recv/recvfrom/recvmsg/recvmmsg path.
Linux TCP/IP recv is not strict in its option parsing: it ignores
flags that it does not expect. Tests verify that this bit in the
flags argument is indeed not accessed.

The intended semantics are obviously not maintained when passing
this flag to Linux sockets.