getsockopt: intercept SO_TIMESTAMPING
Intercept getsockopt `SO_TIMESTAMPING` requests to invert the
setsockopt operation.

This wrapper library emulates support for hardware timestamps on
platforms that do not support it by replacing SOF_.. requests from
HW to SW.

An application that calls getsockopt expects to see the flags that
it passed, not the converted ones. So intercept that call and
invert the translation.