recvmsg: cmsg: convert scm_timestamping sw to hw
Timestamps are returned in a cmsg as struct scm_timestamping, with
fields ts[0] through ts[2].

Software timestamps are returned in ts[0].
Hardware timestamps are returned in ts[2].

Copy ts[0] to ts[2] to return results for hardware timestamp requests,
which are based on software timestamps.

Do this copy unconditionally, these fields are undefined if the given
timestamp type is not requested (i.e., there should be no expectation
that the field is zero).