onload_ordered_epoll_wait: export function
Support processes that call onload_ordered_epoll_wait.

This implementation presents the API but does not implement the
behavior that differentiates the call from standard epoll_wait.

onload_ordered_epoll_wait returns both ready file descriptors (like
epoll) and a timestamp for each. That allows ordering the ready
events. Some file descriptors may not be ordered with respect to
others. For instance for a file descriptor that is not a socket. Then
the behavior falls back to standard epoll. Unordered events are
identified by a zero tv_sec field in their timespec.

This implementation is a trivial wrapper around epoll_wait. It returns
zero for every onload_ordered_epoll_event.

See also the Onload User Guide and the feature test on github at
https://github.com/Xilinx-CNS/onload/blob/master/src/tests/onload/wire_order/wire_order_server.c